requirements.txt
channels==3.0.3
django-cors-headers==3.13.0
djangorestframework====3.13.0
django-model-utils==4.2.0
Django==3.1.5
djangorestframework-jwt==1.11.0